requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.5
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pytz==2020.5
sqlparse==0.4.1
toml==0.10.2
